Authentication/requirements.txt
django==5.0.2
djangorestframework==3.14.0
python-dotenv==1.0.0
djangorestframework-simplejwt==5.3.1
pika==1.3.2 